This pull request should greatly boost page speed (without breaking the site).
Assuming that all JS files are currently only required to make the typeahead inputs
on the index page work, we can make the following optimisation, which moves loading
and parsing out of the entire path to the first DOMContentLoaded event.

We simply require a single async JS file, which moves the file out of the critical
path. The file will then load the remaining scripts, making sure they are loaded
and parsed in order.

Be aware that on slow networks the user might run into cases where the input boxes
are not yet enhanced with TypeScript and won't show autocomplete. I guess this falls
under progressive enhancement.

Admittedly, the current solution is a bit primitive, but it will do for now. When
the page dependencies get more complex, you might want to introduce a proper
dependency loader instead.

Closes #54